% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate}
\alias{aggregate}
\title{Aggregation for Oakland variables.}
\usage{
aggregate(dat, compute)
}
\arguments{
\item{dat}{Data with a column containing census tracts ("tractid10") and variable of interest ("var").}

\item{compute}{"mean", "median", or "sum"}
}
\description{
Take mean, median, or sum of variables by gentrification, ethnoracial, and income grouping.
}
