% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_categories.R
\name{aggregate_categories}
\alias{aggregate_categories}
\title{Aggregation for Oakland variables.}
\usage{
aggregate_categories(dat, compute = "mean")
}
\arguments{
\item{dat}{Data with a column containing census tracts ("tractid10") and variable of interest ("var").}

\item{compute}{"mean", "median", or "sum"}
}
\value{
Aggregated data frame.
}
\description{
Take mean, median, or sum of variables by gentrification, ethnoracial, and income grouping.
}
