% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inmigration_map_panel.R
\name{inmigration_map_panel}
\alias{inmigration_map_panel}
\title{Produce 4x4 map panel of inmigration count in Oakland}
\usage{
inmigration_map_panel(
  data,
  shp_tracts,
  breaks = scales::extended_breaks(n = 6),
  limits = NULL,
  save = F,
  savename = "plot.png",
 
    caption = "\\nSES Ranges by Equifax Risk Scores: Low = missing or <580, Moderate = 580-649, Middle = 650-749, High = 750+\\nHousing Period Ranges: Boom = 2002-2006, Bust = 2007-2009, Recovery = 2010-2014, Post-Recovery = 2015-2017."
)
}
\arguments{
\item{data}{Data with a column containing tractid10, year, ses, and inmigration.}

\item{shp_tracts}{"US_tract_2010.shp" loaded object}

\item{breaks}{Gradient scale breaks, either numeric vector or scales::extended_breaks(n = 6)}

\item{limits}{Gradient scale limits, c(min, max)}

\item{save}{T if user would like to return plot object and save file, F (default) to just return object.}

\item{savename}{File name of map for saving.}

\item{caption}{Figure caption}
}
\value{
4x4 panel maps of inmigration count, by SES and period.
}
\description{
This function takes in data and produces a 4x4 panel of inmigration
counts using a gradient color scale, across SES and periods.
Not intended to be used with aggregated CCP data.
}
