% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_graph.R
\name{line_graph}
\alias{line_graph}
\title{Produce line graph of variable by month}
\usage{
line_graph(
  dat,
  var,
  group = "race",
  save = F,
  savename = "plot.png",
  caption = paste0(frb_caption, ses_caption, period_caption)
)
}
\arguments{
\item{dat}{Data with a column containing variable of interest, fips column, and grouping variable.}

\item{var}{Name of variable to plot, as a string.}

\item{group}{Category for color grouping: "race" (default) or "neighborhood"}

\item{save}{T if user would like to return plot object and save file, F (default) to just return object.}

\item{savename}{File name of map for saving.}

\item{caption}{Caption for figure}
}
\value{
Line graph of variable by month, grouped by race or neighborhood category
}
\description{
This function takes in data and produces a line graph of variables by month,
grouped by neighborhood category or race.
}
