% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{make_map}
\alias{make_map}
\title{Produce Oakland map of continuous variable with gradient color scale}
\usage{
make_map(data, var, savename, labs, scale_label = scales::label_comma())
}
\arguments{
\item{data}{Data with a column containing census tracts and variable of interest.}

\item{var}{Name of column containing variable to plot.}

\item{savename}{File name of map for saving.}

\item{labs}{Adjust map title and axes (must be in the form "labs(title = "TITLE HERE")")}

\item{scale_label}{Adjust scales of gradient key (need to fix)}
}
\value{
Map of variable of interest.
}
\description{
This function takes in data and produces a census tract map of Oakland
representing the variable using a gradient color scale.
}
