% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lollipop plot.R
\name{plot_lollipop}
\alias{plot_lollipop}
\title{Produce lollipop plot by gentrification, ethnoracial, and income category.}
\usage{
plot_lollipop(
  dat,
  limits,
  y_title,
  title,
  savename,
  caption = "Data was aggregated by taking mean of all tracts in a category"
)
}
\arguments{
\item{limits}{Y-axis limits}

\item{y_title}{Y-axis title}

\item{title}{Figure title}

\item{savename}{File name of map for saving.}

\item{data}{Data with a column containing census tracts and variable of interest.}
}
\value{
Lollipop plot of variable by gentrification, ethnoracial, and income category.
}
\description{
This function takes in data and produces a horizontal lollipop plot by
gentrification, ethnoracial, and income category.
}
