% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_stacked_bar.R
\name{race_stacked_bar}
\alias{race_stacked_bar}
\title{Produce stacked bar charts filled by race category}
\usage{
race_stacked_bar(
  dat,
  x_group,
  title = NULL,
  save = F,
  savename = "plot.png",
  caption = paste0(frb_acs_caption, ses_caption)
)
}
\arguments{
\item{dat}{Data with a "variable" column containing races, a "value" column, a "cat" column for faceting labels, and x-axis grouping variable}

\item{x_group}{Category for x-axis grouping}

\item{title}{Plot title}

\item{save}{T if user would like to return plot object and save file, F (default) to just return object.}

\item{savename}{File name of map for saving.}

\item{caption}{Caption for figure}
}
\value{
Panel of stacked bar charts of different groupings
}
\description{
This function takes in data and produces a stacked bar chart, filled by race
category. The facet and x-axis labels can be customized.
}
