% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_bar_move.R
\name{stacked_bar_move}
\alias{stacked_bar_move}
\title{Produce a stacked bar chart for move types}
\usage{
stacked_bar_move(
  dat,
  facet = "ses",
  group = "ethnoracial",
  save = F,
  savename = "plot.png",
 
    caption = "\\nSES Ranges by Equifax Risk Scores: Low = missing or <580, Moderate = 580-649, Middle = 650-749, High = 750+\\nHousing Period Ranges: Boom = 2002-2006, Bust = 2007-2009, Recovery = 2010-2014, Post-Recovery = 2015-2017.\\n"
)
}
\arguments{
\item{dat}{Data with a column containing variables of interest and grouping variables.}

\item{facet}{Name of category to use as facet grouping variable, default: "ses"}

\item{group}{Category for x-axis grouping, can be any column of labels, default: "ethnoracial"}

\item{save}{T if user would like to return plot object and save file, F (default) to just return object.}

\item{savename}{File name of map for saving.}

\item{caption}{Caption for figure}
}
\value{
Stacked bar charts of move types based on different groupings
}
\description{
This function takes in data and produces a stacked bar chart for
different types of moves. Intended to be used with oakViz::aggregate_categories()
}
