% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{make_map}
\alias{make_map}
\title{Produce Oakland map of continuous variable with gradient color scale}
\usage{
make_map(
  data,
  var,
  title = "Title",
  shp_tracts,
  coord = F,
  save = F,
  savename = "plot.png"
)
}
\arguments{
\item{data}{Data with a column containing census tracts and variable of interest.}

\item{var}{Name of column containing variable to plot.}

\item{title}{Figure title}

\item{shp_tracts}{"US_tract_2010.shp" loaded object}

\item{coord}{T if plotting coordinate values (lat, lon). Default is F.}

\item{save}{T if user would like to return plot object and save file, F (default) to just return object.}

\item{savename}{File name of map for saving.}
}
\value{
Map of variable of interest.
}
\description{
This function takes in data and produces a census tract map of Oakland
representing the variable using a gradient color scale. Census tract
column must be named "tractid10".
}
