% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{make_map}
\alias{make_map}
\title{Produce Oakland map of continuous variable with gradient color scale}
\usage{
make_map(
  data,
  var,
  shp_tracts,
  palette = "sequential",
  breaks = scales::extended_breaks(n = 6),
  labels = scales::percent,
  limits = NULL,
  coord = F,
  save = F,
  savename = "plot.png",
 
    caption = "\\nSES Ranges by Equifax Risk Scores: Low = missing or <580, Moderate = 580-649, Middle = 650-749, High = 750+\\nHousing Period Ranges: Boom = 2002-2006, Bust = 2007-2009, Recovery = 2010-2014, Post-Recovery = 2015-2017."
)
}
\arguments{
\item{data}{Data with a column containing census tracts and variable of interest.}

\item{var}{Name of column containing variable to plot.}

\item{shp_tracts}{"US_tract_2010.shp" loaded object}

\item{palette}{Color palette: "sequential" (default) or "diverging"}

\item{breaks}{Gradient scale breaks, numeric vector}

\item{labels}{Gradient scale labels, character vector}

\item{limits}{Manual limits for color scale, numeic vector: c(min, max)}

\item{coord}{T if plotting coordinate values (lat, lon). Default is F.}

\item{save}{T if user would like to return plot object and save file, F (default) to just return object.}

\item{savename}{File name of map for saving.}

\item{caption}{Figure caption}
}
\value{
Map of variable of interest.
}
\description{
This function takes in data and produces a census tract map of Oakland
representing the variable using a gradient color scale. Census tract
column must be named "tractid10".
}
