% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map panel.R
\name{make_map_panel}
\alias{make_map_panel}
\title{Produce Oakland map panel of continuous variable with gradient color scale}
\usage{
make_map_panel(
  data,
  var,
  title = "Title",
  shp_tracts,
  coord = F,
  save = F,
  savename = "plot.png",
  periods
)
}
\arguments{
\item{data}{Data with a column containing census tracts and variable of interest.}

\item{var}{Name of column containing variable to plot.}

\item{title}{Title for panel of maps.}

\item{shp_tracts}{"US_tract_2010.shp" loaded object}

\item{coord}{T if plotting coordinate values (lat, lon). Default is F.}

\item{save}{T if user would like to return plot object and save file, F (default) to just return object.}

\item{savename}{File name of map for saving.}

\item{periods}{Name of column containing four distinct time periods for mapping.}
}
\value{
Map panel of variable of interest across four periods.
}
\description{
This function takes in data and produces a panel of census tract maps of
Oakland representing the variable using a gradient color scale, across
four distinct periods. Should have "tractid10" column for census tracts.
}
