% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period bar chart.R
\name{plot_bar_periods}
\alias{plot_bar_periods}
\title{Produce grouped bar chart across four periods.}
\usage{
plot_bar_periods(
  dat,
  var,
  limits,
  group = "gent",
  compute = "mean",
  title = "Title",
  y_title = "Y-axis title",
  save = F,
  savename = "plot.png",
  caption = "Data was aggregated by taking mean of all tracts in a category"
)
}
\arguments{
\item{dat}{Data with a column containing census tracts and variable of interest.}

\item{var}{Name of column containing variable to plot.}

\item{limits}{Y-axis limits}

\item{group}{Category for x-axis grouping: "gent" (default), "ethnoracial", or "income"}

\item{compute}{Method of summarizing values: "mean" (default) or "median"}

\item{title}{Figure title}

\item{y_title}{Y-axis title}

\item{save}{T if user would like to return plot object and save file, F (default) to just return object.}

\item{savename}{File name of map for saving.}

\item{caption}{Caption for figure}
}
\value{
Grouped bar chart across four periods.
}
\description{
This function takes in data and produces a bar chart grouped by either
gentrification, ethnoracial, or income category across four housing periods.
}
