% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period bar chart.R
\name{plot_bar_periods}
\alias{plot_bar_periods}
\title{Produce grouped bar chart across four periods.}
\usage{
plot_bar_periods(
  dat,
  var,
  limits,
  group = "gent",
  save = F,
  savename = "plot.png",
  caption = paste0(frb_caption, ses_caption, period_caption)
)
}
\arguments{
\item{dat}{Data with a column containing census tracts and variable of interest.}

\item{var}{Name of column containing variable to plot.}

\item{limits}{Y-axis limits}

\item{group}{Category for x-axis grouping: "gent" (default), "ethnoracial", "income", "ses", or "period"}

\item{save}{T if user would like to return plot object and save file, F (default) to just return object.}

\item{savename}{File name of map for saving.}

\item{caption}{Caption for figure}
}
\value{
Grouped bar chart across four periods.
}
\description{
This function takes in data and produces a bar chart grouped by either
gentrification, ethnoracial, income, ses, or period category across four housing periods.
}
