% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lollipop plot.R
\name{plot_lollipop}
\alias{plot_lollipop}
\title{Produce lollipop plot by gentrification, ethnoracial, and income category.}
\usage{
plot_lollipop(
  data,
  var,
  limits,
  x_title = NULL,
  scale_type = "numeric",
  save = F,
  savename = "plot.png",
 
    caption = "\\nSES Ranges by Equifax Risk Scores: Low = missing or <580, Moderate = 580-649, Middle = 650-749, High = 750+\\nHousing Period Ranges: Boom = 2002-2006, Bust = 2007-2009, Recovery = 2010-2014, Post-Recovery = 2015-2017.\\n"
)
}
\arguments{
\item{data}{Data with column for variable of interest.}

\item{var}{Column name of variable of interest.}

\item{limits}{Y-axis limits.}

\item{x_title}{Title to display along x-axis}

\item{scale_type}{Y-axis scale type: "numeric" or "percent"}

\item{save}{T if user would like to return plot object and save file, F (default) to just return object.}

\item{savename}{File name of map for saving.}

\item{caption}{Figure caption}
}
\value{
Lollipop plot of variable by gentrification, ethnoracial, and income category.
}
\description{
This function takes in data and produces a horizontal lollipop plot by
gentrification, ethnoracial, and income category. Input data needs columns for
variable of interest (titled "var") and tract number (titled "tractid10").
Intended to be used after oakViz::aggregate_categories().
}
