% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lollipop plot.R
\name{plot_lollipop}
\alias{plot_lollipop}
\title{Produce lollipop plot by gentrification, ethnoracial, and income category.}
\usage{
plot_lollipop(
  data,
  var,
  limits,
  title = "Title",
  y_title = "Y-axis Title",
  save = F,
  savename = "plot.png",
  caption = "Data was aggregated by taking mean of all tracts in a category"
)
}
\arguments{
\item{data}{Data with column for variable of interest.}

\item{var}{Column name of variable of interest.}

\item{limits}{Y-axis limits.}

\item{title}{Figure title}

\item{y_title}{Y-axis title}

\item{save}{T if user would like to return plot object and save file, F (default) to just return object.}

\item{savename}{File name of map for saving.}

\item{caption}{Figure caption}
}
\value{
Lollipop plot of variable by gentrification, ethnoracial, and income category.
}
\description{
This function takes in data and produces a horizontal lollipop plot by
gentrification, ethnoracial, and income category. Input data needs columns for
variable of interest (titled "var") and tract number (titled "tractid10").
Intended to be used after oakViz::aggregate_categories().
}
