% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_bar.R
\name{stacked_bar}
\alias{stacked_bar}
\title{Produce panel of stacked bar charts}
\usage{
stacked_bar(
  dat,
  fill = "ses",
  facet = c("ethnoracial", "gent", "income"),
  group = "period",
  save = F,
  savename = "plot.png",
 
    caption = "\\nSES Ranges by Equifax Risk Scores: Low = missing or <580, Moderate = 580-649, Middle = 650-749, High = 750+\\nHousing Period Ranges: Boom = 2002-2006, Bust = 2007-2009, Recovery = 2010-2014, Post-Recovery = 2015-2017.\\n"
)
}
\arguments{
\item{dat}{Data with a column containing variables of interest and grouping variables.}

\item{fill}{Name of column containing fill variable, default is ses.}

\item{facet}{Name of category to use as facet grouping variable: "Ethnoracial", "Gentrification", "Income"}

\item{group}{Category for x-axis grouping, can be any column of labels, period (default)}

\item{save}{T if user would like to return plot object and save file, F (default) to just return object.}

\item{savename}{File name of map for saving.}

\item{caption}{Caption for figure}
}
\value{
Panel of stacked bar charts of different groupings
}
\description{
This function takes in data and produces a panel of stacked bar charts,
based on different groupings. Intended to be used with oakViz::aggregate_categories()
}
